Curso IA Modulo 2/HW_C4/hotel-reservation-system/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
redis==5.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# HTTP requests and API clients
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Background tasks and queue
celery==5.3.4
redis==5.0.1

# Email and SMS
sendgrid==6.10.0
twilio==8.10.0
emails==0.6.0

# Payment integrations
stripe==7.8.0
paypalrestsdk==1.13.3

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
python-dateutil==2.8.2
pytz==2023.3

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
factory-boy==3.3.0

# Code quality
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# CORS and middleware
fastapi-cors==0.0.6
slowapi==0.1.9

# Monitoring and health checks
prometheus-client==0.19.0

# File handling
pillow==10.1.0
python-magic==0.4.27

# JSON and data processing
orjson==3.9.10
pandas==2.1.3

# Frontend (if using Streamlit)
streamlit==1.28.1
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0
